# openapi.yaml
openapi: 3.1.0
info:
  title: Waternity API
  description: |
    Waternity is a transparent water sales and settlement platform built on Hedera Hashgraph.
    This API provides endpoints for managing water projects, wells, settlements, and audit trails.
    
    ## Authentication
    The API uses cookie-based authentication with httpOnly JWT tokens. After successful login,
    a secure session cookie (`wty_sess`) is set with 8-hour expiry.
    
    ## Role-Based Access Control
    - **INVESTOR**: Can invest in projects and view settlements
    - **OPERATOR**: Can create wells/projects, manage operations, control meters/valves
    - **AGENT**: Can verify/approve settlements and anchor documents
    - **ADMIN**: Full system access
    
    ## Hedera Integration
    All operations are recorded on Hedera Consensus Service (HCS) for immutable audit trails.
    Links to HashScan explorer and Mirror Node are provided for verification.
  version: 1.0.0
  contact:
    name: Waternity Team
    url: https://waternity.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://waternity.vercel.app
    description: Production server
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: |
        Authenticate user with username and password. Sets httpOnly session cookie on success.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              investor:
                summary: Investor login
                value:
                  username: investor1
                  password: password123
              operator:
                summary: Operator login
                value:
                  username: operator1
                  password: password123
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Session cookie with JWT token
              schema:
                type: string
                example: wty_sess=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Lax; Max-Age=28800; Path=/
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Clears the session cookie
      operationId: logout
      responses:
        '200':
          description: Logout successful
          headers:
            Set-Cookie:
              description: Cleared session cookie
              schema:
                type: string
                example: wty_sess=; HttpOnly; Secure; SameSite=Lax; Max-Age=0; Path=/
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Returns the current authenticated user's profile information
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'

  # HCS Events
...
  /api/audit/reports:
    get:
      tags:
        - Audit
      summary: Get audit reports
      description: |
        Retrieve audit reports for transparency and compliance.
        This endpoint is publicly accessible.
      operationId: getAuditReports
      parameters:
        - name: wellId
          in: query
          description: Filter by well ID
          schema:
            type: string
        - name: fromDate
          in: query
          description: Start date for report period
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: End date for report period
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Audit reports retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditReportsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # Metadata
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: wty_sess
      description: |
        Session-based authentication using httpOnly JWT cookies.
        The cookie is automatically set after successful login and cleared on logout.
        
        Cookie attributes:
        - HttpOnly: true (prevents XSS attacks)
        - Secure: true (HTTPS only in production)
        - SameSite: Lax (CSRF protection)
        - Max-Age: 28800 (8 hours)
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: User's username
          example: "investor1"
        password:
          type: string
          format: password
          description: User's password
          example: "password123"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string
          example: "Login successful"
    UserProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/UserProfile'
    User:
      type: object
      properties:
        id:
          type: string
          example: "1"
        name:
          type: string
          example: "John Investor"
        role:
          $ref: '#/components/schemas/Role'
    UserProfile:
      type: object
      properties:
        id:
          type: string
          example: "1"
        name:
          type: string
          example: "John Investor"
        role:
          $ref: '#/components/schemas/Role'
        issuedAt:
          type: integer
          description: JWT issued at timestamp
          example: 1705315800
        expiresAt:
          type: integer
          description: JWT expiration timestamp
          example: 1705344600
    Role:
      type: string
      enum:
        - INVESTOR
        - OPERATOR
        - AGENT
        - ADMIN
      description: |
        User roles with different permissions:
        - INVESTOR: Can invest in projects and view settlements
        - OPERATOR: Can create wells/projects, manage operations
        - AGENT: Can verify/approve settlements and anchor documents
        - ADMIN: Full system access

    # HCS Event Schemas (referencing JSON Schema 2020-12)
...
    AuditReportsResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/AuditReport'
        count:
          type: integer
          example: 10
    AuditReport:
      type: object
      properties:
        id:
          type: string
          example: "audit_001"
        wellId:
          type: string
          example: "well_001"
        reportType:
          type: string
          enum: [financial, operational, compliance]
          example: "financial"
        period:
          $ref: '#/components/schemas/Period'
        summary:
          type: object
          description: Report summary data
        links:
          $ref: '#/components/schemas/HederaLinks'
        createdAt:
          type: string
          format: date-time

    # Common Schemas
...
    Period:
      type: object
      properties:
        start:
          type: string
          format: date
          example: "2024-01-01"
        end:
          type: string
          format: date
          example: "2024-01-31"
...
    HederaLinks:
      type: object
      properties:
        hashscan:
          type: string
          format: uri
          description: HashScan explorer link
          example: "https://hashscan.io/testnet/topic/0.0.789012"
        mirror:
          type: string
          format: uri
          description: Mirror Node API link
          example: "https://testnet.mirrornode.hedera.com/api/v1/topics/0.0.789012/messages"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                error: "validation_error"
                message: "Request validation failed"
                details:
                  field: "username"
                  issue: "Username is required"
                timestamp: "2024-01-15T10:30:00Z"
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error: "unauthorized"
                message: "Authentication required"
                timestamp: "2024-01-15T10:30:00Z"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal:
              value:
                error: "internal_error"
                message: "An internal error occurred"
                timestamp: "2024-01-15T10:30:00Z"
...